[web] Create ProductSelector instead of form
